feat: add HTTPS cert pinning to ESP32 cloud client + fix iOS build errors
Firmware: WiFiClientSecure with ISRG Root X1 CA cert (Let's Encrypt
root, valid to 2035) for all cloud API calls. Auto-detects HTTP vs
HTTPS from server URL. New ca_cert.h embeds the root cert in PROGMEM.

iOS fixes: cast sinf() Float return to CGFloat in PongGameView and
SharedARView to resolve Xcode 16 strict type checking errors.

Phase 14.5 Hardening is now complete.

+++ ios/Hookbot/Views/PongGameView.swift
@@ -155,7 +155,7 @@ final class PongScene: SKScene {
         smilePath.move(to: CGPoint(x: -4, y: -4))
         for i in 0...8 {
             let t = CGFloat(i) / 8.0
-            smilePath.addLine(to: CGPoint(x: -4 + t * 8, y: -4 + sinf(Float(t) * .pi) * 3))
+            smilePath.addLine(to: CGPoint(x: -4 + t * 8, y: -4 + CGFloat(sinf(Float(t) * .pi)) * 3))
         }
         let smile = SKShapeNode(path: smilePath)
         smile.strokeColor = .black
